guess work

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -9,15 +9,15 @@
  */
 void print_python_list_info(PyObject *p)
 {
-	int size, i = 0;
+	Py_ssize_t size, i = 0;
 
-	printf("[*] Size of the Python List = %d\n", int(PyList_Size(p)));
-	printf("[*] Allocated = %d\n", int(Py_SIZE(p)));
+	printf("[*] Size of the Python List = %u\n", (unsigned int)PyList_Size(p));
+	printf("[*] Allocated = %u\n", (unsigned int)Py_SIZE(p));
 
-	size = int(PyList_Size(p))
+	size = PyList_Size(p)
 	while (i < size)
 	{
-		printf("Element %d: %s\n", i, (char *)(Py_TYPE(p[i])));
+		printf("Element %u: %s\n", (unsigned int)i, (char *)Py_TYPE(p[i]));
 		i++;
 	}
 }
